backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
click==8.3.1
colorama==0.4.6
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fastapi==0.121.3
gunicorn==25.1.0
h11==0.16.0
idna==3.11
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.11.0
sniffio==1.3.1
sqlparse==0.5.4
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.38.0
whitenoise==6.12.0
